Update README.md
I added /bin/ to instructions because /bin/data is in the repo. I was a little confused whether or not to put the untar'd data folder in /example-mnist/bin or just in /example-mnist until I looked in bin and saw that data had the same exact folders in it.

Cheers.